v1.0.2 - Obenseuer 4.04 compat update
Fix issues introduced by Obenseuer update 4.04 (as reported in ticket #1).

Completely pull out visual patches as they are no longer required.
Disable rerail patch as the functionality for this now appears to be in the basegame.  Left in the plugin build but gated behind a setting (default: false) for any on-the-fly testing.